Fixed the path of include files. Fixed the build.gradle dependencies. Fixed cmake flags for debug and release modes. Fixed cmake include directories. Fixed the view parameter in the startRender method. Added resetSampling to shaders. Added some flags to proguard-rules.pro in order to allow debug the native libraries.

+++ jniLibs/Components/src/main/cpp/MobileRT/Shaders/Whitted.cpp
@@ -138,8 +138,7 @@ void Whitted::shade(RGB &rgb, Intersection &intersection, Ray &ray) const {
             //ray.d = ((ray.d*n) + (N*(n*cost1 - sqrt(cost2)))).norm();
             transmissionRay.direction_ = ((ray.direction_ * refractiveIndice) +
                                           (intersection.normal_ * (refractiveIndice * cost1 -
-                                                                   static_cast<float> (std::sqrt(
-                                                                           cost2)))));
+                                                                   std::sqrt(cost2))));
         } else {// reflection direction
             //ray.d = (ray.d + N*(cost1 * 2)).norm();
             transmissionRay.direction_ = (ray.direction_ + intersection.normal_ * (cost1 * 2));
